Methods getToken() and getUsername() has been made more robust
In the old implementation of the getToken() method, it was found that the statement

return currentUser.token;

could cause a runtime error when localStorage is empty (i.e. the currentUser object is null), and the backend was running on a remote web server such as an Azure Web App.

Therefore this statement has been replaced with:

     if (currentUser) {
        return currentUser.token;
      } else {
        return null;
      }

This fixes the issue.